requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-appconf==1.1.0
django-imagekit==5.0.0
djangorestframework==3.16.1
dotenv==0.9.9
grpcio==1.74.0
grpcio-tools==1.74.0
idna==3.10
milvus-lite==2.5.1
numpy==2.3.2
pandas==2.3.2
pilkit==3.0
pillow==11.3.0
protobuf==6.32.0
pycparser==2.22
PyJWT==2.10.1
pymilvus==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
